requirements.txt
ultralytics>=8
groq>=0
Pillow>=10
python-dotenv>=1
requests>=2